+++ packages/devtools_app/test/test_infra/flutter_test_storage.dart
@@ -16,7 +16,7 @@ class FlutterTestStorage implements Storage {
   }
 
   @override
-  Future setValue(String key, String value) async {
+  Future<void> setValue(String key, String value) async {
     values[key] = value;
   }
 }